feat: finish renaming Commands to CommandGroup
