+++ src/mod.ts
@@ -1,4 +1,4 @@
 export { Command } from "./command.ts";
-export * from "./commands.ts";
+export * from "./command-group.ts";
 export * from "./context.ts";
 export type { CommandOptions } from "./types.ts";
